model_image_classification/requirements.txt
# TIMM 图像分类训练工具 - 依赖列表

# 核心依赖
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Web界面
gradio>=4.0.0

# 数据处理
Pillow>=9.0.0
numpy>=1.21.0
pandas>=1.3.0

# 可视化
matplotlib>=3.5.0

# 配置和日志
PyYAML>=6.0

# 可选：加速
# tqdm>=4.64.0
